Added a fix for windowed fullscreen mode
Most all streamers have the game mute when they alt-tab into a different window/monitor. Not only does it typically get muted, but it also becomes very laggy or even still (When in fullscreen). An easy fix for both the muting and laggyness to be fixed is to pop these 2 console commands in your autoexec.